# swagger/Book-mgmt.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Book mgmt",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "variables": {}
    }
  ],
  "paths": {
...
    "/books/unpublish/659b62c40e372023fb2bf8aa": {
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "unpublish",
        "operationId": "unpublish",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/books/published": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "allbooks",
        "operationId": "allbooks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}